Add global state to control cat generation
We didn't have enough space to maneuver and avoid the cat. Now we don't
create a cat on the next tick if we created one during this one.

However, while this spread the cats and looks like it gives more space
in between, I'm still hitting the cats. Maybe it has to do with the size
of the images or the bounding box.

So, a little improvement, but still cannot avoid cats when switching
lanes.